reco_service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
psycopg2-binary==2.9.9
scikit-learn==1.5.2
numpy==2.1.2
pydantic==2.9.2

